+++ cmd/hidden_lake/composite/_daemon/status_hlc.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+watch -c SYSTEMD_COLORS=1 systemctl status -o cat hidden_lake_composite.service